feat(web): remove manifest and related icons
The manifest file and related icons have been removed from the web app.
This includes the deletion of the manifest.ts file and several icon
files such as android-chrome-192x192.png, android-chrome-512x512.png,
and favicon-32x32.png. These changes are part of a cleanup effort to
streamline the assets and configuration of the web application.